Make Making Better
Add a tagline to all all pages, Make Making Better, and make it a link to a placeholder page.

Update other text on the site, the patterns page and the homepage.

Add an icon for twitter

Fix an errant extra </span> in all the project tiles HTML

Refactor some spacing in the layout (so the big mission statement header section which has a background can go beneath the main header without space in-between, and without introducing negative margins for it or something similar)