Move unclear TODO thing off page
It kinda sounded like it was asking *you* to do these things. Which, it would be great if you could, just in the page there; it'd be cool if those were easy little things you could do to try out the system, but the system isn't there... so it's like, what's the deal? Also, outdated. Expounded on the to-dos in the comments.