Check spelling
There's a lot of idiosyncratic spelling which I mainly only want to avoid so I don't have to add it to the allowed list of spellings (1. Since it takes effort to accept the spellings, I might as well change them; 2. I don't want to typo "clearly" as "cleary" and have it accepted.)